prints the first 50 Fibonacci number

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,16 +8,16 @@
 
 int main(void)
 {
-	int fib1  = 0;
-	int fib2  = 1;
-	int fib3;
-	int count = 1;
+	long int fib1  = 0;
+	long int fib2  = 1;
+	long int fib3;
+	long int count = 1;
 
 	while (count <= 50)
 	{
 		fib3 =  fib2 + fib1;
 		count++;
-		printf("%d", fib3);
+		printf("%ld", fib3);
 
 		if (count <= 50)
 			printf(", ");
